Load 3.3.398 into branch/prcs-import-33.
git-svn-id: svn://svn.bearnip.com/ldmud/branch/prcs-import-33@863 56160b6e-f8f8-0310-bd54-bccebfe6a2f2

+++ src/object.c
@@ -7617,7 +7617,8 @@ static int nesting = 0;  /* Used to detect recursive calls */
 #ifdef DEBUG
         if (nesting > 1)
             debug_message("%s: DEBUG: restore_object(%s) nesting %d size %lu\n"
-                         , time_stamp(), name, nesting, (size_t)(st.st_size+1));
+                         , time_stamp(), name, nesting
+                         , (unsigned long)(st.st_size+1));
 #endif
         if (!buff)
         {
